refactor: extract contract-types dialogs and table to components
- Create external dialog components for contract-types CRUD operations
- Add color badge display with color name in uppercase
- Apply same color badge style to positions, work-locations, and departments pages
