+++ src/renderer/src/pages/departments-page.tsx
@@ -30,6 +30,10 @@ import {
   useUpdateDepartment,
 } from "@/hooks";
 
+const getColorName = (color: string) => {
+  return color.replace("bg-", "").replace("-500", "").toUpperCase();
+};
+
 export function DepartmentsPage() {
   const { t } = useTranslation();
   const queryClient = useQueryClient();
@@ -247,7 +251,6 @@ export function DepartmentsPage() {
                       <TableRow className="hover:bg-muted/50" key={department.id}>
                         <TableCell className="px-4">
                           <span className="inline-flex items-center gap-1.5 rounded-md border border-border px-2 py-0.5 font-medium text-xs">
-                            <span className="inline-flex items-center gap-1.5 rounded-md border border-border px-2 py-0.5 font-medium text-xs">
                             {department.code}
                           </span>
                         </TableCell>
@@ -255,7 +258,10 @@ export function DepartmentsPage() {
                           {department.name}
                         </TableCell>
                         <TableCell className="px-4">
-                          <span className={`h-3 w-3 rounded-full ${department.color}`} />
+                          <span className="inline-flex items-center rounded-md border border-border bg-muted/50 px-2 py-0.5 font-medium text-xs">
+                            <span className={`mr-1.5 h-2 w-2 rounded-full ${department.color}`} />
+                            {getColorName(department.color)}
+                          </span>
                         </TableCell>
                         <TableCell className="px-4">
                           {department.isActive ? (
