+++ src/renderer/src/pages/work-locations-page.tsx
@@ -33,6 +33,10 @@ import {
   useWorkLocations,
 } from "@/hooks";
 
+const getColorName = (color: string) => {
+  return color.replace("bg-", "").replace("-500", "").toUpperCase();
+};
+
 export function WorkLocationsPage() {
   const { t } = useTranslation();
   const queryClient = useQueryClient();
@@ -260,7 +264,10 @@ export function WorkLocationsPage() {
                           {location.name}
                         </TableCell>
                         <TableCell className="px-4">
-                          <span className={`h-3 w-3 rounded-full ${location.color}`} />
+                          <span className="inline-flex items-center rounded-md border border-border bg-muted/50 px-2 py-0.5 font-medium text-xs">
+                            <span className={`mr-1.5 h-2 w-2 rounded-full ${location.color}`} />
+                            {getColorName(location.color)}
+                          </span>
                         </TableCell>
                         <TableCell className="px-4">
                           {location.isActive ? (
